Added User model, authentication and authorization

+++ services/cars.js
@@ -36,25 +36,37 @@ async function publishCar(car) {
     await Car.create(car);
 }
 
-async function editCar(id, car) {
-    const data = await Car.findById(id);
+async function editCar(id, owner, data) {
+    const car = await Car.findById(id);
 
-    data.name = car.name.toLowerCase()
-    data.description = car.description
-    data.imageUrl = car.imageUrl
-    data.price = car.price
-    data.accesories = car.accesories
+    if (owner != car.owner) {
+        throw new Error('User is not owner');
+    }
 
-    await data.save();
+    car.name = car.name.toLowerCase()
+    car.description = data.description
+    car.imageUrl = data.imageUrl
+    car.price = data.price
+    car.accesories = data.accesories
+
+    await car.save();
 }
 
-async function deleteById(id) {
+async function deleteById(id, owner) {
+    const car = await Car.findById(id);
+
+    if (owner != car.owner) {
+        throw new Error('User is not owner');
+    }
     await Car.findByIdAndDelete(id);
 }
 
-async function attachAccessory(carId, accessoryId) {
+async function attachAccessory(carId, accessoryId, owner) {
     const car = await Car.findById(carId);
 
+    if (owner != car.owner) {
+        throw new Error('User is not owner');
+    }
     car.accessories.push(accessoryId);
 
     await car.save();
